Fix regression in recurse_submodules (#585)
This should fix #584.

Apparently the submodule subcommand does not support the --worktree argument, and needs to be run from a checked out worktree. By changing the first checkout to use git worktree (https://git-scm.com/docs/git-worktree, available since git 2.6.7) and the later submodules command to run in the new worktree, git submodules seems to be ok.

Verified this locally with git 2.10 (Sept 2016), similar to #421.