testingextended: helper to get test timeout value (#830)
It's useful to set go test's -timeout flag in different contexts, such
as during integration tests or when using verbose debugging or running
in a resource-constrained environment. Keeping such scenarios in mind,
it's additionally often useful to avoid internal timeouts during certain
tests by specifying a larger timeout duration. Unfortunately, this is
prone to error when the `go test` timeout exceeds the internal timeout,
which may not have been the intent of the test author, resulting in a
new, distinct and surprising failure.

Let's teach testingextended to expose `testing`'s global `timeout` flag
value so tests can refer to a duration that is guaranteed to exceed the
deadline of the test.

Prior art: https://github.com/golang/go/issues/34707#issuecomment-539019076