pkg/thread: fix select precedence (#913)
We want context cancellation or timeout to always take precedence over
an unblocked semaphore, so we never start goroutines that can't
successfully complete their work. We're very close to that goal right
now, but there's a timing bug in our logic.

Select chooses pseudorandomly from the unblocked, non-default cases. No
matter what checks we do beforehand, code like this isn't quite correct:

```go
select {
case <-ctx.Done():
  // abort
case <-semaphoreC:
  // work
}
```

Under load, it's likely that a fair bit of time has elapsed between the
previous checks and this select; since the goroutine is blocked on the
channel reads, it gets parked and some other goroutine has a chance to
run. Once this goroutine is unblocked and runs, it's entirely possible
that _both_ the `ctx.Done()` and `semaphoreC` reads will succeed. In
that circumstance, we don't want to choose randomly! Instead, we want
this:

```go
select {
case <-ctx.Done():
  // abort
case <-semaphoreC:
  select {
  case <-ctx.Done():
    // abort
  default:
    // work
  }
}
```

Now, we're guaranteed to do work only if the semaphore is unblocked and
the context is blocked. This is similar to how RWMutexes are used: once
you acquire the write lock, you typically need to re-check that the
write is valid before mutating data.