Add buf registry module commands
These are currently simply mappings from registry repository
to registry module commands. It's the first step towards a
general renaming from:

* module -> snapshot
* repository - > module

As we aim to be clearer about the distinction between git
and BSR concepts.

Keeping in mind that any change like this one is (almost)
a one-way street on our naming strategy.

Address #4334