Add buf registry snapshot commands
These are currently simply mappings from registry commit
to registry snapshot commands. This is part of the renaming
effort of:

* module -> snapshot
* repository - > module

As we aim to be clearer about the distinction between git
and BSR concepts.

Keeping in mind that any change like this one is (almost)
a one-way street on our naming strategy.

This also tweaks some of our docs.

Address #4334