initial pass at buf plugin image creation
Update the `buf plugin push` command to support building a Docker image
from `Dockerfile.plugin` and pushing it to the BSR's registry. This will
create a docker image with the name `<remote>/<owner>/<name>` and a
unique ID for the tag locally, push it to the registry, and then clean
up the local image when complete (so we don't clutter up user's Docker
images).

This initially supports Docker with BuildKit and simple Docker images
where the context only contains the Dockerfile and no other files (no
support for ADD/COPY instructions).