fix connection upgrade for Docker tests (#1250)
Currently, the Docker tests are occassionally hitting a 10s delay on
shutdown when closing the buildkit session. This appears to be due to
incorrect logic when upgrading the connection from HTTP/1.1 to HTTP/2.
The Docker client API doesn't send the required HTTP2-Settings header
and so the h2c handler doesn't upgrade the connection properly.

Update the Docker engine implementation for tests to correctly upgrade
to a H2C connection, fixing the delay on close.

Update some additional tests in this package to support parallel
execution.