allow pushing an existing image with plugin push (#1359)
Currently, calling 'buf alpha plugin push' requires rebuilding a Docker
image from scratch (including replicating context, build args, caching
arguments, and authentication credentials). In some cases, we may wish
to simply re-use an image build with a previous 'docker build' command.

To support this behavior, add an '--image' argument which can refer to
an existing image in the Docker client to be tagged (with the
appropriate BSR tag) and pushed.