Use latest protocompile in `buf format`, fix things up in the formatter (#1427)
Several aspects of how comments are represented in the AST have changed. Reconciling this with the formatter logic was a fairly major/intrusive change. Along the way, the code is now slightly cleaner (and maybe a little easier to maintain for the next change?). And this also fixes some previously unnoticed issues, mostly involving whitespace around comments. It also adds a compact form for simple message and array literals.

The changes to the golden files demonstrate the format tweaks. The change log has a little more detail in exactly what was changed.