Use protocompile as the compiler (#1463)
This replaces the prior compiler (`github.com/jhump/protoreflect/desc/protoparse`) with a new one in `github.com/bufbuild/protocompile`.

Some test proto files had to be updated because they weren't actually legal. (The old compiler did not enforce some constraints that it should have.) This also updates the case for suppressing lint warnings related to RPC request and response types: they must appear as comments on the method, not comments inside parenthesis next to the actual message types. (This is due to changes in the new compiler to more closely match source code info of `protoc`.)