Modules Proto: Push and Download blobs (#1529)
`PushRequest` and `DownloadResponse` now define a way to move modules as
a manifest of the module's files. The content is modeled as set of blobs
with the manifest linking blobs to path names.

These new manifest and blobs fields will eventually be populated and
should be prefered values over the module field. We'll ignore their
values until we can populate them.

This commit adds tests to the module reader since it must receive but
ignore the new fields. I hope it helps clarify that nil values are
expected to be passed into bufmodule.NewModuleForProto.

Co-authored-by: Julian R Figueroa <jfigueroa@buf.build>
Linear: TCN-632