bufcli: add test, storagemem: NewReadWriteBucketWithOptions (#1540)
Coverage for bufcli.ReadModuleWithWorkspacesDisabled
----------------------------------------------------

ReadModuleWithWorkspacesDisabled didn't have coverage and I intend to do
some minor refactoring soon. Now's a great time to document its
behavior.

ReadModuleWithWorkspacesDisabled took a `appflag.Container`, but that's
a very large interface to create and inject. I didn't see any obvious
test fixture to provide it., so instead of building that I reduced
signature to its minimal interface: a logger and an
app.EnvStdinContainer.

IMHO, it's pretty odd to depend on a large interface type from a
different package. I'll leave that yak shave for another day.

storagemem.NewReadWriteBucketWithOptions
------------------------------

It wasn't possible to initialize an in-memory storage bucket like you
could with NewReadBucket. This new constructor produces a read-write
bucket with the functionally of the read-only bucket.

NewReadWriteBucketWithOptions consumes the code in NewReadBucket, making
the later nothing more than a type cast.

This change is in support of testing ReadModuleWithWorkspacesDisabled.