[BSR-721] New hash blobs constructor and validator (#1544)
Changes in this PR:
- Separated `Digest` (and new `Digester`) code into its own `digest.go`
file
- Separated `NewDigestFromBlob` into `module.go` in case we wanna break
that dependency later.
- Separated test files accordingly (manifest/digest/module).
- Exported `DigestType` type, and supported types (just shake256 for
now). It didn't make sense to have it private if it was required to be
used from the consumer in raw string form anyway.
- Created a new `Digester` interface that allows digesting content w/out
the need to do it via a `Manifest`.
- Removed any "file content" concern from the `Manifest`, now it only
cares about "entries" or "pairs" of paths+ and its digests. Meaning it
doesn't have a `hash` and it doesn't receive `AddContent` anymore, just
`AddEntry(path, digest)`. The digests can be built using the
aforementioned `Digester`, in an upper layer that cares about content.
- Added checks to make sure manifest paths have a unique digest.
- Since `Manifest` doesn't care about content anymore, it also doesn't
check digest types. That was moved down to the `Digest` constructors.

This results in a nicer, slimmer exported funcs for `Manifest`, which
only cares about paths and manifests:

![image](https://user-images.githubusercontent.com/4431392/198500219-6d56ab68-c613-424d-a5b4-65ec4a807190.png)

And for `Digest`, we have 3 constructors, all of those using the same
`NewDigestFromBytes` underneath, which errors (for now) if the dtype is
not supported. This allows for future extension just adding a new dtype,
without breaking exported interface.

Two concerns still in my mind:
- It's possible to declare a `Digest{}` w/out going through a
constructor, so not sure if we want to make `Digest` into an interface.
- The `manifest.AddEntry` func checks that the passed digest is
populated/valid at least, but that check feels out of place. I thought
of doing a `func (*Digest) Valid() bool` but it felt even weirder, why
would we allow a Digest to be invalid?
- If a `Digest` is allowed to be constructed outside of the pkg, then
somehow `digest []bytes` and `hexstr string` could diverge? feels both
like two sources of the same state.

Co-authored-by: Tristan Willy <twilly@buf.build>
Co-authored-by: Tristan Willy <twilly@users.noreply.github.com>