Add support for well known types for buf beta convert (#1494)
Adds ability to call `buf beta convert` without specifying an input when
a well known type is specified;

```
$ buf beta convert --type=google.protobuf.Duration --from <(echo \"3600s\")#format=json
> some proto bytes

$ buf beta convert --type=google.protobuf.Duration --from duration.bin
> "3600s"
```

### Input is defaulted to "."
This means that by default calling `buf beta convert
--type=google.protobuf.Duration` will look for
`google.protobuf.Duration` in the local directory first before resolving
to a baked in well known type

Without explicit setting of input to be "." no errors will be reported
on. This is to stop from `buf beta convert` failing to work in any
directory that might have a broken proto file in. This means that:
- `buf beta convert . --type=google.protobuf.Duration` will fail if
`broken.proto` exists, for example but
- `buf beta convert --type=google.protobuf.Duration` will pass because
the error from parsing `"."` will be discarded because the user did not
specifically request for it to be parsed.

Co-authored-by: Elliot Jackson <13633636+ElliotMJackson@users.noreply.github.com>
Co-authored-by: bufdev <bufdev-github@buf.build>