make cyclic import test deterministic (#1551)
Running tests with `-race` would change execution order in a way
that test expectations fail for one test case. The behavior is otherwise
deterministic (so the test is always passing in CI, since we don't run
the tests with the race detector).

This fixes that test case so it passes with or without the race
detector enabled.