push: send a module as a manifest and blobs
Let's oush the module to the registry as plain files represented as a
bag of blobs and a manifest.

This does not remove pushing a protobuf encoded module.

Additionally, the change to deprecate the module field in a push request
makes testing for the field a linting issue. Let's continue testing this
field as it's still in use.