remove default_opts from buf.plugin.yaml
After using buf.plugin.yaml for a bit, we discovered that 'default_opts'
had more potential to confuse users than make their lives easier. It
also was unclear how a user could reset options to no options if a
plugin used 'default_opts'.

As a result, we've migrated to remove all use of 'default_opts' so a
plugin behaves the same when executed locally or remotely in
buf.gen.yaml. We also make use of the 'opts' field under 'registry:' to
set appropriate options for generated code by the Go module proxy or NPM
registries.