Introduce path to start using plain Connect client stubs (#1578)
This will help us address the client-side of BSR-336.

The `connectclient.Config` and the generic `connectclient.Make` function
serve the previous role of `Provider`, in terms of making sure the
client stubs are configured consistently. But they allow us to do so
without extra code generation and to just leverage the code generated by
`protoc-gen-connect-go`.

This change allows us to port over call sites to plain stubs piecemeal,
one package at a time. As an example of how the new stuff is used, this
PR ports over the `bufgen` package.

Co-authored-by: Elliot Jackson <13633636+elliotmjackson@users.noreply.github.com>