[BSR-581] Proto Definition for Set/List Recommended Resources (Plugins/Repos) (#1616)
Added proto definition for two new RPCs in `RecommendationService`.
* `ListRecommendedResources`: accepts empty request body, returns a list of resources,
each of which is either a Repository or a Curated plugin, in the order specified in`SetRecommnededResourcesRequest`. If some resources set by `SetRecommnededResources` 
were deleted or made private after the list was set, they will not be included in the response.
* `SetRecommnededResources`: accepts a list of `SetRecommendedResource`,
each including an owner and a name, and returns an empty response. For
each (owner, name) pair, there has to exist a repository or curated
plugin with that name, and owned by that owner. It also needs to be
public. This will set the list of recommended resources to this list of
resources, in the same order.