Push a minimal set of blobs.
This is an exercise in frustration to push only the set of blobs needed
to build a module. The bug's cause was building a module did not expose
the bucket used to build the module with, hence the manifest consumed
the source bucket containing module-unrelated files.

bufmodulebuild tries to abstract away building but had, and still has,
many rough edges. Some of note are:

 - Build configuration is in the bucket, but the build section of
   configuration is not read from the bucket.

 - Build configuration cannot be written to `buf.yaml`.

 - Incorrect build configuration in a bucket can make a bucket appear
   buildable when it is not.

 - Some build configuration is not in the build section nor in the bucket,
   instead being hard coded as options.

 - BuildForBucket would take options directly, even though it was a
   method. It's a plain function masquerading as a method, causing
   callers to believe they're injecting when really the tail call was
   constructing independent behavior.

I tried to remove passing configuration by value or to serialize
configuration into a bucket if it must be passed as a value (e.g. in
many tests). It turns out writing configuration from configuration
structs is not possible; the build section is omitted. I suspect it's
impossible to round-trip configuration to entities to configuration and
back to identical entities.

Included in this commit is a compromise in exposing a minimal module
bucket. It should produce correct buckets if configuration is passed
that was read from the source bucket and no options are used to change
the configuration's specified behavior. This is the case with `buf push`
but not the case with some other commands, notably `buf lint` given it's
`--config` option.

Changes include:

 - Remove bufcli.ReadModule wrapper on bufmodulebuild.  This wrapper
   didn't provide any value over directly invoking bufmodulebuild.

 - Remove logging support from bucket builder. Logging is not used, and
   should have been passed as a constructor option anyway.

 - Move BuildForBucket to be a function that construct a builder and
   calls a new BuildForBucket method.

 - Type alias ModuleBucketBuilder to avoid rewriting the world if the
   bucket builder needs a new method.

 - Expose the minified bucket by returning a struct with both a usable
   module and the bucket used to create it. This struct could include a
   usable configuration structure to inspect or write as a buf.yaml, but
   I'll leave that as an exercise for another unfortunate soul.

Fixes: TCN-796