check for existing image before pushing to registry (#1663)
Update `buf alpha plugin push` to first check if an existing image
digest exists in the OCI registry. If found, we can re-use it instead of
pushing a new tag of the same image.

This should also greatly improve the performance of pushing images when
the image already exists remotely.