image filtering: don't include entire enclosing message/service (#1659)
If an enclosing message is not part of the transitive dependency graph,
but a nested message therein is, the enclosing message will be stripped
of its fields. This produces a smaller filtered descriptor set but does not
impede the use of the results for dynamic messages or dynamic RPC.

Similarly, if a method is indicated as a type filter, its enclosing service
will be stripped of other unreferenced methods.

Finally, this attempts to fix/clarify the behavior around when custom
options are included since they are also known extensions. If known
extensions are included in the filtered set AND one of the options
message types is part of the transitive dependency graph, all of the
relevant custom options will be included. Otherwise, if custom
options are included, ones actually referenced in options on the
elements in the transitive dependency graph will be present.