Add PR title linter (#1738)
We have had inconsistent PR titles across our various repositories -
some start with capital letters, some don't, some end in periods, some
don't. No one way is the right way, but from day one I did PR titles
with capital letters and no periods at the end, and one way is better
than no way, so @akshayjshah had the idea that we should just enforce
this as part of linting with this action. Trying this on bufbuild/buf,
and it it works, we'll add to our other repositories as well.