Move manifest<>proto mappers to bufapimodule.
The manifest module depended on protobuf when it shouldn't have. These
mappers are the boundary between manifest types and protobuf types,
meaning they should be in a different package in a scope that has
visibility into both.

This places the mappers into bufapimodule since that module encapsulates
the API boundary when transferring module content.