bufwasm: moves more initialization to compile time and fixes leak (#1956)
Before, we forgot to close the cache allocated by the wasm plugin. What
was more subtle was that a large amount of initialization was happening
per `Run` when it was known at compile time. This sorts out those issues
so that as much overhead as possible happens before running a plugin.

The result is pretty sweet, as we are in tens of microseconds per Run.
```bash
$ go test -run='^$' -bench '^Benchmark.*' ./private/bufpkg/bufwasm -count=6 > old.txt
$ go test -run='^$' -bench '^Benchmark.*' ./private/bufpkg/bufwasm -count=6 > new.txt
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/bufbuild/buf/private/bufpkg/bufwasm
             │   old.txt    │              new.txt               │
             │    sec/op    │   sec/op     vs base               │
PluginRun-12   358.48µ ± 2%   23.85µ ± 5%  -93.35% (p=0.002 n=6)
```

Signed-off-by: Adrian Cole <adrian@tetrate.io>