Fix error when installing plugin from zip file (#1991)
When checking for errors, we should have used errors.Is against
storage.ErrClosed instead of os.ErrClosed. The image is closed
automatically when it is uploaded via the http.Client, so we need to
properly guard against double close.

Additionally, add some context to errors so we can easily identify where
they occur. The previous logic would just print "already closed" which
made it difficult to find the call site where the error was occurring.