Update JavaPackage modifier to support java_package
The current version of the JavaPackage modifier doesn't support
prepending a java_package_prefix to an existing value of the
`java_prefix` file option. This makes sense for generation with remote
plugins and managed mode, however for remote packages we may wish to
preserve the user's `java_prefix` setting and use it instead of the
protobuf package name in the final output. This can be useful to provide
more control over the Java packages in the generated output, and also
avoid known issues with Java keywords in package names that prevent
successful compilation.