Warn invalid imports: if not from local or direct dependencies (#2081)
Module `a` depends on `b`, `b` depends on `c`, `a` does not explicitly
depends on `c`, it is a transitive dependency. Module `a` does `import
"file/in/c";` and it works, it shouldn't.

Fix: Make sure all import statements are explicitly declared as
dependency in the `buf.yaml` file. Warn if this is not true.