Resume sync from last sync point (#2170)
There are some important considerations:
- Sync is resumed _after_ the last known commit for a particular branch.
- Repositories that previously called `buf push` can still take
advantage of resumption with `buf sync` if the BSR can find a sync
point. The BSR currently looks for a tag on the BSR commit that _looks
like_ a Git commit hash. In some cases this may not be correct, and
users will have to manually create a new tag with the correct Git commit
hash.
- Rebasing is not supported and currently fails the entire sync in a
partial state.
- Sync points _must_ be encountered, otherwise sync fails after
completion.