Dependencies warning fix: remote discovery for mod update and mod prune (#2178)
This PR is composed of two fixes, one is consequence of the other:

1. Recently we introduced a warning if we detected an import that didn't
come from local target files, or direct dependencies; this includes
workspace modules, and more importantly, transitive dependencies.

This warning shouldn't be present for local unnamed modules, since they
might be local-only, and not required to be pushed to any BSR. So that
is the first fix, removing this warning from local unnamed modules.

2. In the case that one of those local modules depended on remote BSR
modules, and those modules were enterprise (i.e. its remote was
different than "buf.build"), then the commands `buf mod update` and `buf
mod prune` would default to "buf.build" to fetch/clean the dependencies,
which forced users to set a vanity name in their local modules that
matched their remotes, so the command would recognize correctly the
remote and fetch dependencies from the right place.

I'm introducing a `DiscoveryRemote` function that iterates over all
declared direct dependencies in the `buf.yaml` file, and returns the
first remote that's different than "buf.build" if any, so in case there
is a local module that has remote dependencies, we don't force the users
to name such modules, and it can remain being unnamed, which results in
we successfully removing the annoying warning from point 1.

Note: Any named module won't go through that discovery process, and the
remote in the name will be used.