Add buf alpha package <registry>-version commands (#2182)
Resolves a package plugin reference to a remote package version.

Examples:

Get the version of the eliza module and the connect-go plugin for use
with go modules.
```
$ buf alpha package go-version --module=buf.build/bufbuild/eliza --plugin=buf.build/bufbuild/connect-go
        v1.7.0-20230609151053-e682db0d9918.1
```

Use a specific module version and plugin version.
```
    $ buf alpha package go-version --module=buf.build/bufbuild/eliza:e682db0d99184be88b41c4405ea8a417 --plugin=buf.build/bufbuild/connect-go:v1.7.0
        v1.7.0-20230609151053-e682db0d9918.1
```

With different remotes:

```
    $ buf alpha package go-version --module=bufbuild.internal/bufbuild/eliza --plugin=buf.internal/bufbuild/connect-go
WARN    This command is in alpha. It is hidden for a reason. This command is purely for development purposes, and may never even be promoted to beta, do not rely on this command's functionality. To suppress this warning, set BUF_ALPHA_SUPPRESS_WARNINGS=1
Usage:
  buf alpha package go-version --module=<buf.build/owner/repository[:ref]> --plugin=<buf.build/owner/plugin[:version]> [flags]

Flags:
  -h, --help   help for get

Global Flags:
      --debug               Turn on debug logging
      --log-format string   The log format [text,color,json] (default "color")
      --timeout duration    The duration until timing out (default 2m0s)
  -v, --verbose             Turn on verbose mode

Failure: module and plugin must be from the same remote

```