Add `ListRepositoryTagsByReference` rpc (#2274)
This adds a new RPC for `ListRepositoryTagsForReference`.
There is an argument to simply add to the old RPC, but given the way
our APIs have been structured, I think adding a new RPC to address this
need is better than simply adding to the old RPC. Even though it would
still
be backwards compatible, the new RPC provides greater clarity.