Add new branch RPCs and change existing RPCs (#2283)
Adds a new RPC for branches, `GetDefaultBranch`, which returns the
current default
branch mapped to BSR main.

Adds three new fields to `RepositoryCommit`:
- `tag_count` -- number of tags associated with the BSR commit
- `git_commit_metadata_count` -- number of git commits with metadata
associated with the BSR commit
- `git_commit_no_metadata_count` -- number of git commit labels without
metadata associated with the BSR commit

Adds pagination of `ListGitMetadataForReference`.

Adds a new RPC for reference, `ListGitCommitsNoMetadataForReference`,
which returns
a paginated list of git commit labels that do not have git metadata.