GetLabelsInNamespace RPC (#2313)
Cannot use existing `GetLabels` RPC because of the way the label
messages data structure is built. Passing a name means in reality
passing both a name+namespace, which I don't need in this scenario. The
usecase is: Give me labels in namespace "foo", with an optional filter
label names []string.