Sync single branch by default (#2329)
The expected most common use case for the `sync` command is in CI, on
pull requests or individual pushes to remotes. This means we should sync
a single branch by default, whatever one I'm currently checked in. Also
added a flag `--all-branches` to go in the usual sync order and sync all
remote branches.