Implement file options in managed mode v2 (#2344)
This PR 
* moves `Override` from `bufimagemodifyv2` to `bufgenv2`, see
https://github.com/bufbuild/buf/pull/2311#discussion_r1270887980.
* implements the all file options

Notes:
* In `v1`, `java_outer_classname` and `java_multiple_files` [preserve
exisiting values in the proto source
files](https://github.com/bufbuild/buf/pull/2002), meaning that if a
proto file has `option java_outer_classname = xyz;`, managed mode will
not modify this option for this file. This is not added to `v2` yet, and
I don't think we should have this behavior in `v2` because with
`disable` in `v2` users can solve this problem. (For example, generating
for [message `GenericConnectionPoolProto` in
envoy](https://buf.build/envoyproxy/envoy/file/main:envoy/extensions/upstreams/http/generic/v3/generic_connection_pool.proto)
would fail if managed mode sets its source file's `java_outer_classname`
to `GenericConnectionPoolProto` because it collides with the class name)
* `cc_enable_arenas` is still added, although deprecated, because I
think users will less likely to migrate to `v2` if they see diffs on
byte arrays in the generated code. It is added to avoid producing this
diff, because this can affect users who do not generated cpp files at
all. We can omit this option in `v2` docs so that new users won't see it
or use it.
* `java_string_check_utf8` is still added, because it turned out to be
NOT deprecated (I was wrong). For more information, this file option has
no effect in `proto3` but it does in `proto2`.


Next steps:
* add a converter: I manually converted the sample `buf.gen.yaml` in
managed mode section of the docs, and it turned out to require a certain
(non-trivial) level of understanding of both `v1` and `v2` of this
configuration. I think it would make a real difference if there is a
converter.
* add integration tests