Add AttachGitTags RPC (#2354)
New RPC to attach git tags to already synced git commits. There is
another way we could do this, which is having the CLI fetching the BSR
commit id/name from the git sha, and then with that id, use the existing
"TagRepositoryCommit" RPC. But that would mean a couple more network
trips for something the backend can resolve easily.