Travel back from a custom git starting point (#2371)
Add the ability to set a custom starting point when traveling back in
the git tree. So far we have only needed to travel from HEAD of
branches, but to be able to sync tags from recently synced git commits,
we need to travel back from specific detached git commits.