Pass imports to buflint.Handler and filter imports internally (#2391)
Our linters should not consider imports when reporting lint errors. For
almost all of our linters, this means that we do not need access to the
import descriptors when performing lint checks. However, this is not the
case for a couple:

- `PACKAGE_NO_IMPORT_CYCLE` - this should report an import cycle even if
it involves imports. It is a bug that we were not reporting this up
until this point.
- Our upcoming linters for `protovalidate` - some require access to
import descriptors.

This moves the import filtering to below the `buflint.Handler` level.
The vast majority of our linters remain the same, and so do the helpers
they call, however we have one helper, `newFilesWithImportsCheckFunc`,
that passes import `protosource.Files` to the lint checker. When this is
used, errors should still not be reported for imports, however this does
give full access to all files.

Note that this means that `buflint.Handler` and `bufbreaking.Handler`
have different calling requirements, but they were always meant (and
have APIs exposed as) to be independent packages that happened to call
the same backing infrastructure via `internal` packages under the hood.